Add work class
