+++ app/src/main/java/io/github/wulkanowy/services/sync/SyncManager.kt
@@ -2,17 +2,16 @@ package io.github.wulkanowy.services.sync
 
 import android.os.Build.VERSION.SDK_INT
 import android.os.Build.VERSION_CODES.O
-import androidx.work.ExistingWorkPolicy
+import androidx.work.ExistingWorkPolicy.REPLACE
+import androidx.work.OneTimeWorkRequest
 import androidx.work.WorkManager
 import io.github.wulkanowy.data.repositories.preferences.PreferencesRepository
-import io.github.wulkanowy.services.sync.base.WorkerConfiguration
 import io.github.wulkanowy.services.sync.channels.SyncChannel
 import timber.log.Timber
 import javax.inject.Inject
 
 class SyncManager @Inject constructor(
     private val syncChannel: SyncChannel,
-    private val workers: Map<Class<*>, @JvmSuppressWildcards WorkerConfiguration>,
     private val preferencesRepository: PreferencesRepository
 ) {
 
@@ -22,11 +21,7 @@ class SyncManager @Inject constructor(
     }
 
     fun start() {
-        if (preferencesRepository.isServiceEnabled) {
-            workers.forEach {
-                WorkManager.getInstance()
-                    .enqueueUniqueWork(it.key.simpleName, ExistingWorkPolicy.REPLACE, it.value.oneTimeRequest)
-            }
-        }
+        WorkManager.getInstance()
+            .enqueueUniqueWork(SyncWorker::class.java.simpleName, REPLACE, OneTimeWorkRequest.from(SyncWorker::class.java))
     }
 }
