+++ app/src/main/java/io/github/wulkanowy/services/sync/SyncWorker.kt
@@ -1,0 +1,31 @@
+package io.github.wulkanowy.services.sync
+
+import android.content.Context
+import androidx.work.ListenableWorker
+import androidx.work.RxWorker
+import androidx.work.WorkerParameters
+import com.squareup.inject.assisted.Assisted
+import com.squareup.inject.assisted.AssistedInject
+import io.github.wulkanowy.services.sync.works.Work
+import io.reactivex.Completable
+import io.reactivex.Single
+
+class SyncWorker @AssistedInject constructor(
+    @Assisted appContext: Context,
+    @Assisted workerParameters: WorkerParameters,
+    private val works: Set<@JvmSuppressWildcards Work>
+) : RxWorker(appContext, workerParameters) {
+
+    override fun createWork(): Single<Result> {
+        return Completable.mergeDelayError(works.map { it.create() })
+            .toSingleDefault(Result.success())
+            .onErrorReturn { Result.failure() }
+    }
+
+    @AssistedInject.Factory
+    interface Factory {
+
+        fun create(appContext: Context, workerParameters: WorkerParameters): ListenableWorker
+    }
+}
+
