--- app/src/main/java/io/github/wulkanowy/services/sync/base/WorkerConfiguration.kt
@@ -1,11 +1,0 @@
-package io.github.wulkanowy.services.sync.base
-
-import androidx.work.OneTimeWorkRequest
-import androidx.work.PeriodicWorkRequest
-
-interface WorkerConfiguration {
-
-    val periodicRequest: PeriodicWorkRequest
-
-    val oneTimeRequest: OneTimeWorkRequest
-}
