--- app/src/main/java/io/github/wulkanowy/services/sync/factory/InjectableWorkerFactory.kt
@@ -1,10 +1,0 @@
-package io.github.wulkanowy.services.sync.factory
-
-import android.content.Context
-import androidx.work.ListenableWorker
-import androidx.work.WorkerParameters
-
-interface InjectableWorkerFactory {
-
-    fun create(appContext: Context, workerParameters: WorkerParameters): ListenableWorker
-}
