+++ app/src/main/java/io/github/wulkanowy/services/sync/works/Work.kt
@@ -1,0 +1,9 @@
+package io.github.wulkanowy.services.sync.works
+
+import io.reactivex.Completable
+
+interface Work {
+
+    fun create(): Completable
+}
+
